requirements.txt
flask
pdfminer.six
python-docx
spacy
gunicorn